# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VenueCloud API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/notes:
    get:
      operationId: listEventNotes
      tags: [notes]
      summary: List notes for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
  /events/{id}/tasks:
    get:
      operationId: listEventTasks
      tags: [tasks]
      summary: List tasks for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  # Functions
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all event leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Accounts
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: relatedType
          in: query
          schema:
            type: string
        - name: relatedId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Notes
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: relatedType
          in: query
          schema:
            type: string
        - name: relatedId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Appointments
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports
      parameters:
        - name: reportType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/run:
    get:
      operationId: runReport
      tags: [reports]
      summary: Run a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"

  # Calendar
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        leadName:
          type: string
        division:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        leadType:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        probability:
          type: ["integer", "null"]
        leadStatus:
          type: string
          enum: [New, Tentative, Definite, Cancelled]
        decisionDate:
          type: ["string", "null"]
        lastActivity:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        salesperson:
          type: ["string", "null"]
        referralType:
          type: ["string", "null"]
        functionType:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        estimatedAttendance:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        primaryContactId:
          type: ["integer", "null"]
        billingContactId:
          type: ["integer", "null"]
        primaryContactName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - leadName
        - leadStatus
        - createdAt
        - updatedAt
    LeadInput:
      type: object
      properties:
        leadName:
          type: string
        division:
          type: string
        site:
          type: string
        leadType:
          type: string
        budget:
          type: number
        probability:
          type: integer
        leadStatus:
          type: string
        decisionDate:
          type: string
        nextAction:
          type: string
        owner:
          type: string
        salesperson:
          type: string
        functionType:
          type: string
        eventDate:
          type: string
        estimatedAttendance:
          type: integer
        description:
          type: string
        primaryContactId:
          type: integer
      required:
        - leadName
    LeadUpdate:
      type: object
      properties:
        leadName:
          type: string
        division:
          type: string
        leadStatus:
          type: string
        budget:
          type: number
        probability:
          type: integer
        decisionDate:
          type: string
        nextAction:
          type: string
        owner:
          type: string
        salesperson:
          type: string
        estimatedAttendance:
          type: integer
        description:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        priority:
          type: string
          enum: [High, Medium, Low]
        description:
          type: ["string", "null"]
        relatedType:
          type: ["string", "null"]
        relatedId:
          type: ["integer", "null"]
        contactId:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        status:
          type: string
          enum: [Open, Closed]
        salesperson:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        relatedName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - priority
        - status
        - createdAt
        - updatedAt
    TaskInput:
      type: object
      properties:
        name:
          type: string
        priority:
          type: string
        description:
          type: string
        relatedType:
          type: string
        relatedId:
          type: integer
        contactId:
          type: integer
        category:
          type: string
        status:
          type: string
        salesperson:
          type: string
        dueDate:
          type: string
        site:
          type: string
      required:
        - name
        - priority
    TaskUpdate:
      type: object
      properties:
        name:
          type: string
        priority:
          type: string
        description:
          type: string
        status:
          type: string
        category:
          type: string
        result:
          type: string
        salesperson:
          type: string
        dueDate:
          type: string
    Note:
      type: object
      properties:
        id:
          type: integer
        relatedType:
          type: ["string", "null"]
        relatedId:
          type: ["integer", "null"]
        noteText:
          type: string
        salesperson:
          type: ["string", "null"]
        internal:
          type: ["boolean", "null"]
        createdAt:
          type: string
      required:
        - id
        - noteText
        - createdAt
    NoteInput:
      type: object
      properties:
        relatedType:
          type: string
        relatedId:
          type: integer
        noteText:
          type: string
        salesperson:
          type: string
        internal:
          type: boolean
      required:
        - noteText
    NoteUpdate:
      type: object
      properties:
        noteText:
          type: string
        salesperson:
          type: string
        internal:
          type: boolean
...
    Report:
      type: object
      properties:
        id:
          type: integer
        reportName:
          type: string
        description:
          type: ["string", "null"]
        reportType:
          type: ["string", "null"]
        folder:
          type: ["string", "null"]
        sqlQuery:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - reportName
        - createdAt
    ReportInput:
      type: object
      properties:
        reportName:
          type: string
        description:
          type: string
        reportType:
          type: string
        folder:
          type: string
        sqlQuery:
          type: string
        createdBy:
          type: string
      required:
        - reportName
    ReportResult:
      type: object
      properties:
        reportId:
          type: integer
        reportName:
          type: string
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        rowCount:
          type: integer
      required:
        - reportId
        - reportName
        - columns
        - rows
        - rowCount
...